requirements.txt
streamlit
openai
pillow
reportlab
